# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Timewise database management system",
        "title": "timewise-dbms",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/dbms/v1/schedule_participant": {
            "get": {
                "description": "Get all schedule participants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Get all schedule participants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TwScheduleParticipant"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create schedule participant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Create schedule participant",
                "parameters": [
                    {
                        "description": "Schedule participant object",
                        "name": "schedule_participant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                }
            }
        },
        "/dbms/v1/schedule_participant/schedule/{scheduleId}": {
            "get": {
                "description": "Get schedule participants by schedule ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Get schedule participants by schedule ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schedule_participant_dtos.ScheduleParticipantInfo"
                            }
                        }
                    }
                }
            }
        },
        "/dbms/v1/schedule_participant/workspace/{workspaceId}/schedule/{scheduleId}": {
            "get": {
                "description": "Get schedule participants by schedule ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Get schedule participants by schedule ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspaceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schedule_participant_dtos.ScheduleParticipantInfo"
                            }
                        }
                    }
                }
            }
        },
        "/dbms/v1/schedule_participant/{id}": {
            "get": {
                "description": "Get schedule participant by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Get schedule participant by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule Participant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                }
            },
            "put": {
                "description": "Update schedule participant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Update schedule participant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule Participant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule participant object",
                        "name": "schedule_participant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete schedule participant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Delete schedule participant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule Participant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TwScheduleParticipant"
                        }
                    }
                }
            }
        },
        "/dbms/v1/schedule_participant/{scheduleId}/participants": {
            "get": {
                "description": "Check if any workspace user in the list is a participant in the schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule_participant"
                ],
                "summary": "Check if workspace user is participant in the schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "scheduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "List of workspace user IDs (comma-separated)",
                        "name": "workspace_user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is participant",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "404": {
                        "description": "User is not participant",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.TwBoardColumn": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/models.TwWorkspace"
                },
                "workspace_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwSchedule": {
            "type": "object",
            "properties": {
                "all_day": {
                    "type": "boolean"
                },
                "boardColumn": {
                    "$ref": "#/definitions/models.TwBoardColumn"
                },
                "board_column_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "extra_data": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "priority": {
                    "type": "string"
                },
                "recurrence_pattern": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_transcript": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                },
                "workspace": {
                    "$ref": "#/definitions/models.TwWorkspace"
                },
                "workspaceUser": {
                    "$ref": "#/definitions/models.TwWorkspaceUser"
                },
                "workspace_id": {
                    "type": "integer"
                },
                "workspace_user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwScheduleParticipant": {
            "type": "object",
            "properties": {
                "assignByUser": {
                    "$ref": "#/definitions/models.TwWorkspaceUser"
                },
                "assign_at": {
                    "type": "string"
                },
                "assign_by": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invitation_sent_at": {
                    "type": "string"
                },
                "invitation_status": {
                    "type": "string"
                },
                "response_time": {
                    "type": "string"
                },
                "schedule": {
                    "$ref": "#/definitions/models.TwSchedule"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspaceUser": {
                    "$ref": "#/definitions/models.TwWorkspaceUser"
                },
                "workspace_user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwUserEmail": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_linked_to": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.TwUser"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TwWorkspaceUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "extra_data": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "userEmail": {
                    "$ref": "#/definitions/models.TwUserEmail"
                },
                "user_email_id": {
                    "type": "integer"
                },
                "workspace": {
                    "$ref": "#/definitions/models.TwWorkspace"
                },
                "workspace_id": {
                    "type": "integer"
                },
                "workspace_key": {
                    "type": "string"
                }
            }
        },
        "schedule_participant_dtos.ScheduleParticipantInfo": {
            "type": "object",
            "properties": {
                "assign_at": {
                    "type": "string"
                },
                "assign_by": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invitation_sent_at": {
                    "type": "string"
                },
                "invitation_status": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "profile_picture": {
                    "type": "string"
                },
                "response_time": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "status_workspace_user": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "workspace_user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}